+++ packages/react-devtools-shared/src/constants.js
@@ -30,6 +30,9 @@ export const SESSION_STORAGE_RELOAD_AND_PROFILE_KEY =
 export const LOCAL_STORAGE_SHOULD_PATCH_CONSOLE_KEY =
   'React::DevTools::appendComponentStack';
 
+export const LOCAL_STORAGE_TRACE_UPDATES_ENABLED_KEY =
+  'React::DevTools::traceUpdatesEnabled';
+
 export const PROFILER_EXPORT_VERSION = 4;
 
 export const CHANGE_LOG_URL =