+++ packages/react-dom/src/events/ChangeEventPlugin.js
@@ -262,10 +262,10 @@ const ChangeEventPlugin = {
 
   extractEvents: function(
     topLevelType,
-    eventSystemFlags,
     targetInst,
     nativeEvent,
     nativeEventTarget,
+    eventSystemFlags,
   ) {
     const targetNode = targetInst ? getNodeFromInstance(targetInst) : window;