+++ packages/react-interactions/accessibility/src/__tests__/FocusTable-test.internal.js
@@ -8,6 +8,7 @@
  */
 
 import {createEventTarget} from 'react-interactions/events/src/dom/testing-library';
+import {emulateBrowserTab} from '../shared/emulateBrowserTab';
 
 let React;
 let ReactFeatureFlags;
@@ -45,8 +46,12 @@ describe('FocusTable', () => {
         TabbableScope,
       );
 
-      return ({onKeyboardOut, id, wrap}) => (
-        <FocusTable onKeyboardOut={onKeyboardOut} id={id} wrap={wrap}>
+      return ({onKeyboardOut, id, wrap, allowModifiers}) => (
+        <FocusTable
+          onKeyboardOut={onKeyboardOut}
+          id={id}
+          wrap={wrap}
+          allowModifiers={allowModifiers}>
           <table>
             <tbody>
               <FocusTableRow>
@@ -112,6 +117,20 @@ describe('FocusTable', () => {
       );
     }
 
+    it('handles keyboard arrow operations with allowModifiers', () => {
+      const Test = createFocusTableComponent();
+
+      ReactDOM.render(<Test allowModifiers={true} />, container);
+      const buttons = document.querySelectorAll('button');
+      const a1 = createEventTarget(buttons[0]);
+      a1.focus();
+      a1.keydown({
+        key: 'ArrowRight',
+        altKey: true,
+      });
+      expect(document.activeElement.textContent).toBe('A2');
+    });
+
     it('handles keyboard arrow operations', () => {
       const Test = createFocusTableComponent();
 
@@ -130,7 +149,7 @@ describe('FocusTable', () => {
       });
       expect(document.activeElement.textContent).toBe('B2');
 
-      const b2 = createEventTarget(document.activeElement);
+      let b2 = createEventTarget(document.activeElement);
       b2.keydown({
         key: 'ArrowLeft',
       });
@@ -150,6 +169,14 @@ describe('FocusTable', () => {
       c1.keydown({
         key: 'ArrowUp',
       });
+      expect(document.activeElement.textContent).toBe('B1');
+      // Should be a no-op due to modifier
+      b2 = createEventTarget(document.activeElement);
+      b2.keydown({
+        key: 'ArrowUp',
+        altKey: true,
+      });
+      b2 = createEventTarget(document.activeElement);
       expect(document.activeElement.textContent).toBe('B1');
     });
 
@@ -356,6 +383,69 @@ describe('FocusTable', () => {
         key: 'ArrowLeft',
       });
       expect(document.activeElement.textContent).toBe('A3');
+    });
+
+    it('handles keyboard arrow operations mixed with tabbing', () => {
+      const [FocusTable, FocusRow, FocusCell] = createFocusTable(TabbableScope);
+      const beforeRef = React.createRef();
+      const afterRef = React.createRef();
+
+      function Test() {
+        return (
+          <>
+            <input placeholder="Before" ref={beforeRef} />
+            <FocusTable tabScope={TabbableScope}>
+              <div>
+                <FocusRow>
+                  <FocusCell>
+                    <input placeholder="A1" />
+                  </FocusCell>
+                  <FocusCell>
+                    <input placeholder="B1" />
+                  </FocusCell>
+                  <FocusCell>
+                    <input placeholder="C1" />
+                  </FocusCell>
+                </FocusRow>
+              </div>
+              <div>
+                <FocusRow>
+                  <FocusCell>
+                    <input placeholder="A2" />
+                  </FocusCell>
+                  <FocusCell>
+                    <input placeholder="B2" />
+                  </FocusCell>
+                  <FocusCell>
+                    <input placeholder="C1" />
+                  </FocusCell>
+                </FocusRow>
+              </div>
+            </FocusTable>
+            <input placeholder="After" ref={afterRef} />
+          </>
+        );
+      }
+
+      ReactDOM.render(<Test />, container);
+      beforeRef.current.focus();
+
+      expect(document.activeElement.placeholder).toBe('Before');
+      emulateBrowserTab();
+      expect(document.activeElement.placeholder).toBe('A1');
+      emulateBrowserTab();
+      expect(document.activeElement.placeholder).toBe('After');
+      emulateBrowserTab(true);
+      expect(document.activeElement.placeholder).toBe('A1');
+      const a1 = createEventTarget(document.activeElement);
+      a1.keydown({
+        key: 'ArrowRight',
+      });
+      expect(document.activeElement.placeholder).toBe('B1');
+      emulateBrowserTab();
+      expect(document.activeElement.placeholder).toBe('After');
+      emulateBrowserTab(true);
+      expect(document.activeElement.placeholder).toBe('B1');
     });
   });
 });