+++ scripts/rollup/bundles.js
@@ -675,6 +675,20 @@ const bundles = [
       'react',
       'react-interactions/events/keyboard',
       'react-interactions/accessibility/tabbable-scope',
+      'react-interactions/accessibility/focus-control',
+    ],
+  },
+
+  {
+    bundleTypes: [NODE_DEV, NODE_PROD, FB_WWW_DEV, FB_WWW_PROD],
+    moduleType: NON_FIBER_RENDERER,
+    entry: 'react-interactions/accessibility/focus-contain',
+    global: 'ReactFocusContain',
+    externals: [
+      'react',
+      'react-interactions/events/focus',
+      'react-interactions/events/keyboard',
+      'react-interactions/accessibility/focus-manager',
     ],
   },
 
@@ -683,20 +697,6 @@ const bundles = [
     moduleType: NON_FIBER_RENDERER,
     entry: 'react-interactions/accessibility/focus-manager',
     global: 'ReactFocusManager',
-    externals: [
-      'react',
-      'react-interactions/events/keyboard',
-      'react-interactions/events/focus',
-      'react-interactions/accessibility/tabbable-scope',
-      'react-interactions/accessibility/focus-control',
-    ],
-  },
-
-  {
-    bundleTypes: [NODE_DEV, NODE_PROD, FB_WWW_DEV, FB_WWW_PROD],
-    moduleType: NON_FIBER_RENDERER,
-    entry: 'react-interactions/accessibility/focus-control',
-    global: 'ReactFocusControl',
     externals: ['react'],
   },
 
@@ -711,12 +711,13 @@ const bundles = [
   {
     bundleTypes: [NODE_DEV, NODE_PROD, FB_WWW_DEV, FB_WWW_PROD],
     moduleType: NON_FIBER_RENDERER,
-    entry: 'react-interactions/accessibility/focus-list',
-    global: 'ReactFocusList',
+    entry: 'react-interactions/accessibility/focus-group',
+    global: 'ReactFocusGroup',
     externals: [
       'react',
       'react-interactions/events/keyboard',
       'react-interactions/accessibility/tabbable-scope',
+      'react-interactions/accessibility/focus-manager',
     ],
   },
 ];
@@ -726,7 +727,7 @@ const fbBundleExternalsMap = {
   'react-interactions/events/keyboard': 'ReactEventsKeyboard',
   'react-interactions/events/tap': 'ReactEventsTap',
   'react-interactions/accessibility/tabbable-scope': 'ReactTabbableScope',
-  'react-interactions/accessibility/focus-control': 'ReactFocusControl',
+  'react-interactions/accessibility/focus-manager': 'ReactFocusManager',
 };
 
 // Based on deep-freeze by substack (public domain)